Add flow libdef header file
Function overloading is only possible from `declare` statements creating
an intersection type for the function, not by providing default values
for generic parameters.

A generic parameter may be referenced and even mutated in the function,
but a new value for it may not be created. Setting a default parameter
value is considered creation.

So default class filters must move from the type signature into the
function body with the header file declaring the overloaded forms of
each query function.